v2.2: Re-introduce topography in IGI version.
Following the approach of MK's original anaglyfo function:
- do not include topography in electrode coordinates,
- read topography from separate input file and correct electrode and
  node z-coordinates for topography.

Note that topography is accounted for in the modelling only. The inverse
problem is still treated in a flat model, not corrected for topography.
A consequence of this that we should be aware of is that spatial deriva-
tives used for Tikhonov regularization are not computed in the x- and z-
directions anymore, but parallel and perpendicularly to the ground sur-
face, such that the induced smoothing is not strictly isotropic anymore.

Francois Lavoue, 24 August 2016

-----------
  On branch dev_IGI_FL
  Changes to be committed:
 	modified:   00README.txt
 	modified:   forward_parameters.m
 	modified:   src/src_ip4di/create_mesh3.m
 	renamed:    src/src_nogui_ip4di/plot_model.m -> src/src_nogui_ip4di/old-src/plot_model.m
 	modified:   src/src_nogui_ip4di/plot_model_on_forward_mesh.m